# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/policy": {
      "post": {
        "operationId": "PolicyController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "tags": [
          "Policy"
        ]
      },
      "get": {
        "operationId": "PolicyController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 5,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by policy category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search keyword for name or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "id",
              "type": "string",
              "enum": [
                "id",
                "name",
                "rating",
                "premium",
                "popularity"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all policies",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PolicyResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "tags": [
          "Policy"
        ]
      }
    },
    "/policy/{id}": {
      "get": {
        "operationId": "PolicyController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get policy with signed URLs",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PolicyResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "tags": [
          "Policy"
        ]
      },
      "patch": {
        "operationId": "PolicyController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update policy",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PolicyResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "tags": [
          "Policy"
        ]
      },
      "delete": {
        "operationId": "PolicyController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remove policy",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PolicyResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "tags": [
          "Policy"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your Supabase access token"
      }
    },
    "schemas": {
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "CreatePolicyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Comprehensive Health Coverage"
          },
          "category": {
            "type": "string",
            "example": "health"
          },
          "provider": {
            "type": "string",
            "example": "HealthSecure"
          },
          "coverage": {
            "type": "number",
            "example": 100000
          },
          "premium": {
            "type": "string",
            "example": "0.8 ETH/month"
          },
          "rating": {
            "type": "number",
            "example": 0
          },
          "popular": {
            "type": "boolean",
            "example": false
          },
          "features": {
            "example": [
              "Emergency Care",
              "Prescription Drugs",
              "Mental Health"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "example": "Complete healthcare coverage with blockchain-verified claims processing"
          },
          "files": {
            "description": "Files to upload for the policy",
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        },
        "required": [
          "name",
          "category",
          "provider",
          "coverage",
          "premium",
          "rating",
          "popular",
          "features"
        ]
      },
      "PolicyDocumentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "policy_id": {
            "type": "number"
          },
          "signedUrl": {
            "type": "string",
            "description": "Signed Supabase URL for accessing this document"
          }
        },
        "required": [
          "id",
          "name",
          "policy_id",
          "signedUrl"
        ]
      },
      "PolicyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "coverage": {
            "type": "number"
          },
          "premium": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "popular": {
            "type": "boolean"
          },
          "description": {
            "type": "object"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "policy_documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyDocumentResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "provider",
          "coverage",
          "premium",
          "rating",
          "popular",
          "features",
          "policy_documents"
        ]
      },
      "UpdatePolicyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Comprehensive Health Coverage"
          },
          "category": {
            "type": "string",
            "example": "health"
          },
          "provider": {
            "type": "string",
            "example": "HealthSecure"
          },
          "coverage": {
            "type": "number",
            "example": 100000
          },
          "premium": {
            "type": "string",
            "example": "0.8 ETH/month"
          },
          "rating": {
            "type": "number",
            "example": 0
          },
          "popular": {
            "type": "boolean",
            "example": false
          },
          "features": {
            "example": [
              "Emergency Care",
              "Prescription Drugs",
              "Mental Health"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "example": "Complete healthcare coverage with blockchain-verified claims processing"
          },
          "files": {
            "description": "Files to upload for the policy",
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
...
    }
  }
}